requirements.txt
opencv-python
numpy
pyserial
picamera2